Revert "Move  to NotificationWithStatus from NotificationGroupWithStatus"
This reverts commit c9dd090e7b1799d7e0ae939af5470f7217727f37.
