+++ app/javascript/mastodon/features/notifications_v2/components/notification_status.tsx
@@ -3,7 +3,7 @@ import { FormattedMessage } from 'react-intl';
 import NotificationsActiveIcon from '@/material-icons/400-24px/notifications_active-fill.svg?react';
 import type { NotificationGroupStatus } from 'mastodon/models/notification_group';
 
-import type { LabelRenderer } from './notification_with_status';
+import type { LabelRenderer } from './notification_group_with_status';
 import { NotificationWithStatus } from './notification_with_status';
 
 const labelRenderer: LabelRenderer = (values) => (
